task07/requirements.txt
hello-agents[all]>=0.2.0
requests
beautifulsoup4
python-dotenv
huggingface_hub
gradio>=4.0.0
